feat: 🚀 Enhanced Employees List and Updated Vacations Container
In this commit, significant changes were made to the EmployeesList and DateRangeVacationsContainer components of the application. The details are as follows:

In the EmployeesList component, several new features were added:
The ability to select an employee from the list has been added.
The selected employee is visually highlighted in the list.
An employee's hiring date is now displayed in the list.
Pagination controls have been added to the list.
The DateRangeVacations component has been renamed to DateRangeVacationsContainer.
A utility function, formatDate, has been created to format date strings in the dd/MM/yyyy format.